+++ internal/collector/tracker.go
@@ -29,6 +29,10 @@ const (
 	QueryJobsLogEntries  = "jobs_log.entries"
 	QueryJobsLogGroups   = "jobs_log.groups"
 	QueryStatsEnabled    = "jobs_log.stats_enabled"
+	QueryChangesCoverage = "changes.coverage"
+	QueryChangesLog      = "changes.log"
+	QueryChangesCluster  = "changes.cluster_settings"
+	QueryChangesTables   = "changes.table_settings"
 	QuerySnapshotRepos   = "snapshots.repositories"
 	QuerySnapshots       = "snapshots.list"
 	QueryHeartbeat       = cratedb.QueryLabelHeartbeat
@@ -99,6 +103,10 @@ func NewQueryTracker(cfg map[string]config.CollectorConfig, connCfg config.Conne
 		{QueryJobsLogEntries, "jobs_log", ci("jobs_log")},
 		{QueryJobsLogGroups, "jobs_log", ci("jobs_log")},
 		{QueryStatsEnabled, "jobs_log", 0},
+		{QueryChangesCoverage, "changes", ci("changes")},
+		{QueryChangesLog, "changes", ci("changes")},
+		{QueryChangesCluster, "changes", ci("changes")},
+		{QueryChangesTables, "changes", 0},
 		{QuerySnapshotRepos, "snapshots", ci("snapshots")},
 		{QuerySnapshots, "snapshots", ci("snapshots")},
 		{QueryHeartbeat, "registry", connCfg.HeartbeatInterval.Duration},